pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "immich-ml-metal"
version = "0.1.0"
description = "Metal/ANE-optimized drop-in replacement for Immich's ML service"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]
keywords = ["immich", "machine-learning", "apple-silicon", "mlx", "coreml"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "orjson>=3.9.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "opencv-python>=4.9.0",
    "pyobjc-framework-Vision>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-CoreML>=10.0",
    "mlx>=0.4.0",
    "huggingface-hub>=0.20.0",
    "onnxruntime-silicon>=1.16.0",
    "insightface>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/immich-ml-metal"
Repository = "https://github.com/yourusername/immich-ml-metal"

[project.scripts]
immich-ml-metal = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]